feat: user cannot update or delete other users now

+++ app/src/main/java/hexlet/code/app/service/UserService.java
@@ -3,9 +3,11 @@ package hexlet.code.app.service;
 import hexlet.code.app.dto.UserCreateDTO;
 import hexlet.code.app.dto.UserDTO;
 import hexlet.code.app.dto.UserUpdateDTO;
+import hexlet.code.app.exception.AccessDeniedException;
 import hexlet.code.app.exception.ResourceNotFoundException;
 import hexlet.code.app.mapper.UserMapper;
 import hexlet.code.app.repository.UserRepository;
+import hexlet.code.app.util.UserUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,6 +21,9 @@ public class UserService {
 
     @Autowired
     private UserMapper userMapper;
+
+    @Autowired
+    private UserUtils userUtils;
 
     public List<UserDTO> getAll() {
         var users = userRepository.findAll();
@@ -42,12 +47,25 @@ public class UserService {
     public UserDTO update(UserUpdateDTO dto, Long id) {
         var user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found"));
-        userMapper.update(dto, user);
-        userRepository.save(user);
-        return userMapper.map(user);
+
+        var currentUser = userUtils.getCurrentUser();
+
+        if (currentUser == null || !currentUser.getId().equals(id)) {
+            throw new AccessDeniedException("Access denied");
+        } else {
+            userMapper.update(dto, user);
+            userRepository.save(user);
+            return userMapper.map(user);
+        }
     }
 
     public void delete(Long id) {
-        userRepository.deleteById(id);
+        var currentUser = userUtils.getCurrentUser();
+
+        if (currentUser == null || !currentUser.getId().equals(id)) {
+            throw new AccessDeniedException("Access denied");
+        } else {
+            userRepository.deleteById(id);
+        }
     }
 }
